requirements.txt
click==8.1.3
defusedxml==0.7.1
Flask==2.1.3
fpdf2==2.5.5
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
Werkzeug==2.1.2
zipp==3.8.1
